WebStorage/tests/approved/storage_local_key.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Web Storage</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>storage_local_key</h3>
	</p>
	
	<p>
		<h3>
			<div id="result" >
				<span id="resultMsg"><span class="fail">FAIL<span></span>
			</div>
		</h3>
	</p>

	<hr>

	<div id="log"></div>

	<script>

		if (('localStorage' in window) && window.localStorage !== null){
			localStorage.clear();

			localStorage.setItem("name", "user1");
			localStorage.setItem("age", "20");
			localStorage.setItem("a", "1");
			localStorage.setItem("b", "2");

			try {
				var keys = ["name","age","a","b"];
				function checkMatch(index){
					for (var i = 0; i < keys.length ; i++){
						if(localStorage.key(index) == keys[i]){
							delete keys[i];
							return true;
						}
					}
					return false;
				}
				
				test(function()
				{
					assert_true(checkMatch(0) && checkMatch(1) && checkMatch(2) && checkMatch(3), "key method is failed.");
				}, "key method test")
					
			} catch (e) {
				test(function()
				{
					assert_unreached("key method is failed.");
				}, "key method test")
			}
		} else {
			test(function()
			{
				assert_unreached("localStorage is not supported.");
			}, "localStorage support test")
		}	
	</script>

 </body>
</html>
